Refactoring; added tests

+++ bmlrp.h
@@ -3,10 +3,10 @@
 
 using namespace std;
 
-typedef uint64_t Addr;
+typedef unsigned long long Addr;
 const Addr FirstBit = ((Addr)-1 >> 1) + 1;
 
 //Graph NextLevel(const Graph& clGraph, const vector<Addr>& addrs);
-Graph GetLevel(const Graph& level0, const vector<Addr>& addrs, int level);
+Graph GetLevel(const Graph& level0, const vector<Addr>& addrs, uint level);
 
 #endif // BMLRP_H
